@@ -3,12 +3,12 @@ import 'package:equatable/equatable.dart';
 import 'failures/validation_failure.dart';
 
 abstract class Entity extends Equatable {
-  final String id;
+  final String? id;
 
   Entity(this.id);
 
   @override
-  List<Object> get props => [this.id];
+  List<Object?> get props => [this.id];
 
   /// Validate the entity
   /// @return all errors from validating, empty list if there are no validation errors
@@ -18,7 +18,7 @@ abstract class Entity extends Equatable {
     return errors;
   }
 
-  static validateRequired(final String field, final ValidationTypes type, final List<ValidationInfo> errors) {
+  static validateRequired(final String? field, final ValidationTypes type, final List<ValidationInfo> errors) {
     if (field == null || field.isEmpty) {
       errors.add(new ValidationInfo(type: type));
     }
