@@ -1,11 +1,11 @@
 import 'failure.dart';
 
 class ServerFailure extends Failure {
-  final ServerFailureTypes type;
+  final ServerFailureTypes? type;
 
   ServerFailure({this.type = ServerFailureTypes.unknown});
 
-  List<Object> get props => [...super.props, type];
+  List<Object?> get props => [...super.props, type];
 }
 
 enum ServerFailureTypes {
