@@ -7,9 +7,9 @@ import '../../domain/repositories/piece_repository.dart';
 import '../datasources/piece_local_data_source.dart';
 
 class PieceRepositoryImpl implements PieceRepository {
-  final PieceLocalDataSource localDataSource;
+  final PieceLocalDataSource? localDataSource;
 
-  PieceRepositoryImpl({@required this.localDataSource});
+  PieceRepositoryImpl({required this.localDataSource});
 
   @override
   Future<PieceEntity> addPiece(PieceEntity piece) async {
@@ -18,16 +18,16 @@ class PieceRepositoryImpl implements PieceRepository {
     }
 
     try {
-      return await localDataSource.add(piece);
+      return await localDataSource!.add(piece);
     } on LocalServerException {
       throw ServerFailure();
     }
   }
 
   @override
-  Future<String> removePiece(String id) async {
+  Future<String?> removePiece(String? id) async {
     try {
-      return await localDataSource.remove(id);
+      return await localDataSource!.remove(id);
     } on LocalServerException {
       throw ServerFailure();
     }
@@ -36,7 +36,7 @@ class PieceRepositoryImpl implements PieceRepository {
   @override
   Future<PieceEntity> updatePiece(PieceEntity piece) async {
     try {
-      return await localDataSource.update(piece);
+      return await localDataSource!.update(piece);
     } on LocalServerException {
       throw ServerFailure();
     }
@@ -45,7 +45,7 @@ class PieceRepositoryImpl implements PieceRepository {
   @override
   Future<List<PieceEntity>> getPieces() async {
     try {
-      return await localDataSource.getAll();
+      return await localDataSource!.getAll();
     } on LocalServerException {
       throw ServerFailure();
     }
