@@ -6,14 +6,14 @@ import '../../../../core/use_case.dart';
 import '../repositories/piece_repository.dart';
 
 class PieceAdd extends UseCase<PieceEntity, PieceEntity> {
-  final PieceRepository repository;
+  final PieceRepository? repository;
 
   PieceAdd(this.repository);
 
   @override
   Future<Either<Failure, PieceEntity>> call(PieceEntity piece) async {
     try {
-      return Right(await repository.addPiece(piece));
+      return Right(await repository!.addPiece(piece));
     } on ServerFailure catch (e) {
       return Left(e);
     }
