@@ -10,8 +10,8 @@ class MockLocalDataSource extends Mock implements PieceLocalDataSource {}
 
 void main() {
   group("PieceRepositoryImpl (#repository #cold) ->", () {
-    PieceRepositoryImpl repository;
-    MockLocalDataSource mockLocalDataSource;
+    late PieceRepositoryImpl repository;
+    MockLocalDataSource? mockLocalDataSource;
     final piece = PieceEntity(id: null, title: "Test name", composer: "Composer");
 
     setUp(() {
@@ -22,20 +22,20 @@ void main() {
     group("addPiece()", () {
       test("should return the local and updated piece when adding a piece", () async {
         final piece = PieceEntity(id: null, title: "Test name", composer: "Composer");
-        when(mockLocalDataSource.add(any)).thenAnswer((_) async => piece);
+        when(mockLocalDataSource!.add(any!)).thenAnswer((_) async => piece);
 
         final result = await repository.addPiece(piece);
 
-        verify(mockLocalDataSource.add(piece));
+        verify(mockLocalDataSource!.add(piece));
         verifyNoMoreInteractions(mockLocalDataSource);
         expect(result, isA<PieceEntity>());
         expect(result, equals(PieceEntity(id: piece.id, title: piece.title, composer: piece.composer)));
       });
 
       test("should throw a ServerFailure when the repository throws a LocalServerException", () {
-        when(mockLocalDataSource.add(any)).thenAnswer((_) async => throw LocalServerException());
+        when(mockLocalDataSource!.add(any!)).thenAnswer(((_) async => throw LocalServerException()) as Future<PieceEntity> Function(Invocation));
         expect(() => repository.addPiece(piece), throwsA(isA<ServerFailure>()));
-        verify(mockLocalDataSource.add(piece));
+        verify(mockLocalDataSource!.add(piece));
         verifyNoMoreInteractions(mockLocalDataSource);
       });
 
@@ -48,35 +48,35 @@ void main() {
 
     group("removePiece()", () {
       test("should return the id of the removed piece when removing a piece", () async {
-        when(mockLocalDataSource.remove(any)).thenAnswer((_) async => piece.id);
+        when(mockLocalDataSource!.remove(any)).thenAnswer((_) async => piece.id);
         final result = await repository.removePiece(piece.id);
-        verify(mockLocalDataSource.remove(piece.id));
+        verify(mockLocalDataSource!.remove(piece.id));
         verifyNoMoreInteractions(mockLocalDataSource);
         expect(result, equals(piece.id));
       });
 
       test("should throw a ServerFailure when the repository throws a LocalServerException", () {
-        when(mockLocalDataSource.remove(any)).thenAnswer((_) async => throw LocalServerException());
+        when(mockLocalDataSource!.remove(any)).thenAnswer((_) async => throw LocalServerException());
         expect(() => repository.removePiece(piece.id), throwsA(isA<ServerFailure>()));
-        verify(mockLocalDataSource.remove(piece.id));
+        verify(mockLocalDataSource!.remove(piece.id));
         verifyNoMoreInteractions(mockLocalDataSource);
       });
     });
 
     group("updatePiece()", () {
       test("should return the local and updated piece when updating a piece", () async {
-        when(mockLocalDataSource.update(any)).thenAnswer((_) async => piece);
+        when(mockLocalDataSource!.update(any!)).thenAnswer((_) async => piece);
         final result = await repository.updatePiece(piece);
-        verify(mockLocalDataSource.update(piece));
+        verify(mockLocalDataSource!.update(piece));
         verifyNoMoreInteractions(mockLocalDataSource);
         expect(result, isA<PieceEntity>());
         expect(result, equals(piece));
       });
 
       test("should throw a ServerFailure when the repository throws a LocalServerException", () {
-        when(mockLocalDataSource.update(any)).thenAnswer((_) async => throw LocalServerException());
+        when(mockLocalDataSource!.update(any!)).thenAnswer(((_) async => throw LocalServerException()) as Future<PieceEntity> Function(Invocation));
         expect(() => repository.updatePiece(piece), throwsA(isA<ServerFailure>()));
-        verify(mockLocalDataSource.update(piece));
+        verify(mockLocalDataSource!.update(piece));
         verifyNoMoreInteractions(mockLocalDataSource);
       });
     });
@@ -92,17 +92,17 @@ void main() {
         ];
 
         for (final test in testData) {
-          when(mockLocalDataSource.getAll()).thenAnswer((_) async => test);
+          when(mockLocalDataSource!.getAll()).thenAnswer((_) async => test);
           final result = await repository.getPieces();
-          verify(mockLocalDataSource.getAll());
+          verify(mockLocalDataSource!.getAll());
           expect(result, test);
         }
       });
 
       test("should throw a ServerFailure when the repository throws a LocalServerException", () {
-        when(mockLocalDataSource.getAll()).thenAnswer((_) => throw LocalServerException());
+        when(mockLocalDataSource!.getAll()).thenAnswer(((_) => throw LocalServerException()) as Future<List<PieceEntity>> Function(Invocation));
         expect(() => repository.getPieces(), throwsA(isA<ServerFailure>()));
-        verify(mockLocalDataSource.getAll());
+        verify(mockLocalDataSource!.getAll());
         verifyNoMoreInteractions(mockLocalDataSource);
       });
     });
