@@ -9,8 +9,8 @@ class _MockPieceRepository extends Mock implements PieceRepository {}
 
 void main() {
   group("Edit Piece should (#usecase #cold) ->", () {
-    PieceEdit usecase;
-    _MockPieceRepository mockPieceRepository;
+    late PieceEdit usecase;
+    _MockPieceRepository? mockPieceRepository;
 
     setUp(() {
       mockPieceRepository = _MockPieceRepository();
@@ -19,12 +19,12 @@ void main() {
 
     test('Call the repository with the piece and get the piece back', () async {
       final piece = fakerPiece();
-      when(mockPieceRepository.updatePiece(any)).thenAnswer((_) async => piece);
+      when(mockPieceRepository!.updatePiece(any!)).thenAnswer((_) async => piece);
 
-      final result = await usecase(piece);
+      final Either<Failure, PieceEntity>? result = await usecase(piece);
 
       expect(result, Right(piece));
-      verify(mockPieceRepository.updatePiece(piece));
+      verify(mockPieceRepository!.updatePiece(piece));
       verifyNoMoreInteractions(mockPieceRepository);
     });
   });
