@@ -5,5 +5,5 @@ abstract class Failure extends Equatable {
   bool get stringify => true;
 
   @override
-  List<Object> get props => [];
+  List<Object?> get props => [];
 }
