@@ -8,9 +8,9 @@ part of 'piece_entity.dart';
 
 PieceEntity _$PieceEntityFromJson(Map<String, dynamic> json) {
   return PieceEntity(
-    id: json['id'] as String,
-    title: json['title'] as String,
-    composer: json['composer'] as String,
+    id: json['id'] as String?,
+    title: json['title'] as String?,
+    composer: json['composer'] as String?,
     lastPracticed: json['lastPracticed'] == null
         ? null
         : DateTime.parse(json['lastPracticed'] as String),
